ewb: introduce
It is problematic to put the options found in EMERGE_WORLD_OPTS into
emerge's EMERGE_DEFAULT_OPTS since for builds other than world builds
(revdep-rebuild, single emerges, etc.), most of them are hindersome.
Specifying them on the command line every time a world build is
attempted is a nuisance, which is where this script comes to the remedy.
When called for world builds, it combines EMERGE_DEFAULT_OPTS and
EMERGE_WORLD_OPTS.

Signed-off-by: bedhanger <bedhanger@gmx.de>