.gitconfig: ensure --skip is present when filter-process or smudge are called
"git lfs install --skip-smudge" should be the command that gets called early on when you set up
LFS generally (i.e. outside of a repository).  This will prevent large files from being downloaded
(and instead place little pointers in their stead) when cloning.

It is somewhat strange why this is not the default, as the behaviour described above is the sales
pitch for LFS to begin with.

When you do not skip, the only way to get the pointers is by using git-lfs-clone (instead of just
git-clone).  Now, git-lfs-clone is deprecated...

Signed-off-by: Ann T Ropea <bedhanger@gmx.de>